requirements-prod.txt
torch
torchvision
transformers
faiss-cpu
pillow
numpy
sqlalchemy
python-dotenv
rembg[cpu]